refactor(const): use camelcase instead of caps
