+++ src/components/Timer/Selections.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import './index.css';
-import { TimerState } from 'const/Timer';
+import { TimerState } from 'const/timer';
 import AppBar from '@material-ui/core/AppBar';
 import Tabs from '@material-ui/core/Tabs';
 import Tab from '@material-ui/core/Tab';
